removed basic endpoint, and refactored buddiesRepo position
