--- src/main/java/com/team701/buddymatcher/services/BasicService.java
@@ -1,8 +1,0 @@
-package com.team701.buddymatcher.services;
-
-/**
- * A basic interface for a BasicService
- */
-public interface BasicService {
-    String helloWorld(String name);
-}
